koruby_precise: wasm ビルドの警告を全部潰す (-Wall -Wextra で 0)
- fiber/thread の trampoline: CTX 分割/復元の << 32 を uint64_t 経由に
  (uintptr_t が 32bit の wasm では UB だった)。
- range/array の min/max/minmax 委譲: VALUE* 引数に KORB_NIL を渡していた
  3 箇所を NULL に。
- getpid の deprecation は -D_WASI_EMULATED_GETPID (定義は wasi_missing.c のまま)。
- astro_code_store の version 入出力を PRIx64/SCNx64 に (long が 32bit)。
- astro_build: strip 名の文字列リテラル比較をやめる。
- main.c: prelude_mtime を __wasi__ で切る (unused)。

+++ sample/koruby_precise/builtins/fiber.c
@@ -27,7 +27,7 @@ korb_raise_fiber_error(CTX *c, VALUE *slots, const char *msg)
 static void
 korb_fiber_trampoline(unsigned hi, unsigned lo)
 {
-    CTX *const c = (CTX *)(((uintptr_t)hi << 32) | (uintptr_t)lo);
+    CTX *const c = (CTX *)(uintptr_t)(((uint64_t)hi << 32) | (uint64_t)lo);
     KorbFiberRep *const rep = c->vm->starting_fiber;
     rep->fstate = 1;                                  /* running */
     VALUE arg = rep->transfer;                        /* value from the first resume */
@@ -119,7 +119,7 @@ korb_fiber_switch_in(CTX *c, VALUE *slots, KorbFiberRep *const rep, VALUE xfer, 
         ((ucontext_t *)rep->uctx)->uc_link = NULL;
         c->vm->starting_fiber = rep;
         makecontext((ucontext_t *)rep->uctx, (void (*)(void))korb_fiber_trampoline, 2,
-                    (unsigned)((uintptr_t)c >> 32), (unsigned)((uintptr_t)c & 0xFFFFFFFFu));
+                    (unsigned)((uint64_t)(uintptr_t)c >> 32), (unsigned)((uintptr_t)c & 0xFFFFFFFFu));
     }
     /* activate the fiber's stacks */
     c->slots = rep->vslots; c->slots_top = rep->vslots_top;
